+++ src/lib/components/ColorScales/index.js
@@ -5,9 +5,4 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-
-import React from "react";
-import ReactDOM from "react-dom";
-import App from "./App";
-
-ReactDOM.render(<App />, document.getElementById("root"));
+export { default } from "./ColorScales";
