--- src/lib/components/Select.css
@@ -1,6 +1,0 @@
-.webviz-config-select {
-    width: 100%;
-    overflow-y: auto;
-    font-family: inherit;
-    font-size: inherit;
-}
