+++ src/lib/components/Select/Select.css
@@ -5,9 +5,9 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-
-import React from "react";
-import ReactDOM from "react-dom";
-import App from "./App";
-
-ReactDOM.render(<App />, document.getElementById("root"));
+.webviz-config-select {
+    width: 100%;
+    overflow-y: auto;
+    font-family: inherit;
+    font-size: inherit;
+}
