+++ src/lib/components/WebvizPluginPlaceholder/webviz_plugin_component.css
@@ -1,3 +1,11 @@
+/**
+ * Copyright (c) 2021- Equinor ASA
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
+
 .webviz-config-plugin-wrapper {
     width: 100%;
     box-sizing: border-box;
