+++ src/lib/index.js
@@ -1,7 +1,14 @@
+/**
+ * Copyright (c) 2021- Equinor ASA
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
 /* eslint-disable import/prefer-default-export */
-import WebvizPluginPlaceholder from "./components/WebvizPluginPlaceholder.react";
-import ColorScales from "./components/ColorScales.react";
-import Select from "./components/Select.react";
+import WebvizPluginPlaceholder from "./components/WebvizPluginPlaceholder";
+import ColorScales from "./components/ColorScales";
+import Select from "./components/Select";
 
 import "./flexbox.css"
