+++ tests/test_graph.py
@@ -1,3 +1,12 @@
+##################################################################
+#
+# Copyright (c) 2021- Equinor ASA
+#
+# This source code is licensed under the MIT license found in the
+# LICENSE file in the root directory of this source tree.
+#
+##################################################################
+
 import dash
 import dash_html_components as html
 import webviz_core_components
